% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_schema.R
\name{create_categorical_rule}
\alias{create_categorical_rule}
\title{Create a rule for categorical variables.}
\usage{
create_categorical_rule(name, values)
}
\arguments{
\item{name}{str Name of variable rule is being created for.}

\item{enumerate}{list List of possible values a variable can take.}
}
\description{
Create a rule for categorical variables.
}
\examples{

create_categorical_rule("arbitrary_binary", c("Y", "N"))
}
