% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_schema.R
\name{create_length_rule}
\alias{create_length_rule}
\title{Create a maximum length rule for string variables.}
\usage{
create_length_rule(name, var_length)
}
\arguments{
\item{name}{str Name of variable rule pertains to.}

\item{var_length}{int Number of characters in a string variable.}
}
\description{
Ensures the length of a string variable matches the specified length.
}
\examples{

create_length_rule("arbitrary_id", 20)
}
