% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_schema.R
\name{create_min_max_rule}
\alias{create_min_max_rule}
\title{Create a min/max rule for string variables.}
\usage{
create_min_max_rule(name, min_length, max_length)
}
\arguments{
\item{name}{Str Name of variable rule pertains to.}

\item{min_length}{int Minimum number of characters in a string variable.}

\item{max_length}{int Maximum number of charactesr in a string variable.}
}
\description{
Ensures the length of a string is greater than the minimum and less than the maxium.
}
\examples{

create_min_max_rule("forename", min_length = 1, max_length = 20)
}
