% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_schema.R
\name{create_pattern_rule}
\alias{create_pattern_rule}
\title{Create a pattern rule for variables.}
\usage{
create_pattern_rule(name, pattern)
}
\arguments{
\item{name}{str Name of variable rule is being created for.}

\item{x}{str Regular expression rule.}
}
\description{
Create a pattern rule for variables.
}
\examples{

create_pattern_rule("arbitrary_variable", "[0-9]*")
}
