% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_schema.R
\name{create_simple_rule}
\alias{create_simple_rule}
\title{Create a simple rule from the extracted information conditional on what values are present.}
\usage{
create_simple_rule(
  name,
  min_length,
  max_length,
  length,
  min_inclusive,
  max_inclusive,
  enumeration_values,
  pattern_values
)
}
\arguments{
\item{name}{str Name of variable from XML Schema}

\item{min_length}{int Minimum length of variable, may be NA.}

\item{max_length}{int Maximum length of variable, may be NA.}

\item{length}{int Length of variable, may be NA.}

\item{min_inclusive}{int Minimum value (inclusive) a value may take.}

\item{max_inclusive}{int Maximum value (inclusive) a value may take.}

\item{enumerations}{list List of possible enumerations for a categorical variable.}

\item{patterns}{list List of possible patterns for a variable.}
}
\description{
Create a simple rule from the extracted information conditional on what values are present.
}
