% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_schema.R
\name{create_simple_ruleset}
\alias{create_simple_ruleset}
\title{Create a simple ruleset conditional on the values for a given variable}
\usage{
create_simple_ruleset(x)
}
\arguments{
\item{x}{List A list of lists extracted}
}
\description{
Create a simple ruleset conditional on the values for a given variable
}
\examples{
types <- extract_structure("path/to/xml_schema.xsd")
simple_info <- extract_info_simple(types$simple)
create_simple_ruleset(simple_info)
}
