% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_schema.R
\name{extract_info_complex}
\alias{extract_info_complex}
\title{Extract information from complex data types.}
\usage{
extract_info_complex(x)
}
\arguments{
\item{x}{list List of simple XML schema extracted using extract_structure}
}
\description{
Extract information from complex data types.
}
\examples{

types <- extract_structure("path/to/xml_schema.xsd")
simple_info <- extract_info_simple(types$complex)
names(simple_info)
}
