% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_schema.R
\name{extract_structure}
\alias{extract_structure}
\title{Extract Types}
\usage{
extract_structure(xml_file)
}
\arguments{
\item{xml_file}{XML file to be parsed.}

\item{type}{Type of element to extract from the schema. TRUD defines the following options which are selected by the
noted code.}
}
\description{
Extract different types of data elements from the XML Schema.
}
\examples{
types <- extract_structure("data/xml/ecds/ecdsxml/CDS_011_Emergency_Care-V4-0.xsd")
}
