% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_schema.R
\name{load_xml}
\alias{load_xml}
\title{Load XML Schema}
\usage{
load_xml(xml_file)
}
\arguments{
\item{file_path}{Path to XML schema file.}
}
\description{
Load XML Schema
}
\examples{

xml_schema <- load_xml(xml_file = "~/path/to/schema.xml")
}
