% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_schema.R
\name{write_rules_to_yaml}
\alias{write_rules_to_yaml}
\title{Write ruleset to YAML file.}
\usage{
write_rules_to_yaml(x, output_dir, yaml_file)
}
\arguments{
\item{x}{list ruleset created by create_simple_ruleset() or create_complex_ruleset().}

\item{output_dir}{str Directory to save YAML file to.}

\item{yaml_file}{str Name of YAML file rules are to be written to under output_dir.}
}
\description{
If the specified directory path doesn't exist it will be created.
}
