% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_validation.R
\name{xml_2_rule_set}
\alias{xml_2_rule_set}
\title{Convert XML Schema to rule_setnames}
\usage{
xml_2_rule_set(x)
}
\description{
Takes an XML Schema and convrets it to a rule_set for validating data/variables.
}
